+++ src/Troschuetz.Random/Generators/AbstractGenerator.cs
@@ -55,20 +55,6 @@ namespace Troschuetz.Random.Generators
         #region Constants
 
         /// <summary>
-        ///   Represents the multiplier that computes a double-precision floating point number
-        ///   greater than or equal to 0.0 and less than 1.0 when it gets applied to a nonnegative
-        ///   32-bit signed integer.
-        /// </summary>
-        protected const double IntToDoubleMultiplier = 1.0 / (int.MaxValue + 1.0);
-
-        /// <summary>
-        ///   Represents the multiplier that computes a double-precision floating point number
-        ///   greater than or equal to 0.0 and less than 1.0 when it gets applied to a 32-bit
-        ///   unsigned integer.
-        /// </summary>
-        protected const double UIntToDoubleMultiplier = 1.0 / (uint.MaxValue + 1.0);
-
-        /// <summary>
         ///   The number of left shifts required to transform a 64-bit unsigned integer into a
         ///   nonnegative 32-bit signed integer.
         /// </summary>
@@ -507,7 +493,23 @@ namespace Troschuetz.Random.Generators
         ///   A double-precision floating point number greater than or equal to 0.0, and less than
         ///   1.0; that is, the range of return values includes 0.0 but not 1.0.
         /// </returns>
-        protected static unsafe double ConvertULongToDouble(ulong value)
+        protected static unsafe double ToDouble(uint value)
+        {
+            const double UIntToDoubleMultiplier = 1.0 / (uint.MaxValue + 1.0);
+            return value * UIntToDoubleMultiplier;
+        }
+
+        /// <summary>
+        ///   Converts given <paramref name="value"/> into a double-precision floating point number
+        ///   greater than or equal to 0.0, and less than 1.0; that is, the range of return values
+        ///   includes 0.0 but not 1.0.
+        /// </summary>
+        /// <param name="value">The value that needs to be converted.</param>
+        /// <returns>
+        ///   A double-precision floating point number greater than or equal to 0.0, and less than
+        ///   1.0; that is, the range of return values includes 0.0 but not 1.0.
+        /// </returns>
+        protected static unsafe double ToDouble(ulong value)
         {
             // 12 bits are discarded because double mantissa is 52 bits long. The exponent on the
             // right side generates double values in following range: [1.0, 2.0).