+++ src/Troschuetz.Random/Generators/NR3Generator.cs
@@ -155,7 +155,7 @@ namespace Troschuetz.Random.Generators
 
             // Conversion method uses 52 of the 64 bits available. Therefore, we need to regenerate
             // the whole set and we cannot use the "_bytesAvailable" flag.
-            var result = ConvertULongToDouble((_x + _v) ^ _w);
+            var result = ToDouble((_x + _v) ^ _w);
 
             // Postconditions
             Debug.Assert(result >= 0.0 && result < 1.0);