+++ src/Troschuetz.Random/Generators/XorShift128Generator.cs
@@ -191,7 +191,7 @@ namespace Troschuetz.Random.Generators
 
             // Conversion method uses 52 of the 64 bits available. Therefore, we need to regenerate
             // the whole set and we cannot use the "_bytesAvailable" flag.
-            var result = ConvertULongToDouble(tx + ty);
+            var result = ToDouble(tx + ty);
 
             // Postconditions
             Debug.Assert(result >= 0.0 && result < 1.0);